requirements.txt
absl-py==2.4.0
altair==6.0.0
astunparse==1.6.3
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cryptography==46.0.4
cycler==0.12.1
deepface==0.0.98
faiss-cpu==1.7.4
filelock==3.20.3
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.2
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.7.0
gdown==5.2.1
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-auth-oauthlib==1.2.4
google-pasta==0.2.0
grpcio==1.76.0
gunicorn==25.0.1
h5py==3.15.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.12.1
kiwisolver==1.4.9
libclang==18.1.1
lightdsa==0.0.3
lightecc==0.0.4
lightphe==0.0.20
lz4==4.4.5
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.32
ml_dtypes==0.5.4
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
narwhals==2.16.0
numpy==1.26.4
oauthlib==3.3.1
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
optree==0.18.0
packaging==26.0
pandas==2.3.3
pillow==12.1.0
protobuf==6.33.5
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
smmap==5.0.2
sounddevice==0.5.5
soupsieve==2.8.3
streamlit==1.54.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.3.0
tf_keras==2.20.1
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
watchdog==6.0.0
Werkzeug==3.1.5
wrapt==1.14.2
